requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
psycopg2-binary>=2.9.0
rich>=13.0.0
